Migrated to more stable infrastructure
- webpack
- single component per file
- prettier
