+++ .gitignore
@@ -62,3 +62,6 @@ typings/
 
 # temporary uploads dir
 uploads
+
+# dist files
+public/js
