+++ .prettierrc.json
@@ -1,0 +1,9 @@
+{
+    "tabWidth": 4,
+    "singleQuote": true,
+    "endOfLine": "lf",
+    "trailingComma": "es5",
+    "semi": true,
+    "printWidth": 80
+  }
+
