+++ package.json
@@ -3,10 +3,12 @@
   "version": "0.0.0",
   "private": true,
   "scripts": {
-    "start": "./bin/www",
-    "start:dev": "nodemon ./bin/www"
+    "start": "webpack --mode=production && ./bin/www",
+    "start:dev": "webpack --mode=development --watch",
+    "build": "webpack --mode=production"
   },
   "dependencies": {
+    "axios": "^0.18.0",
     "debug": "~2.6.9",
     "express": "~4.16.0",
     "gm": "^1.23.1",
@@ -18,10 +20,15 @@
     "multer": "^1.4.1",
     "react": "^16.8.1",
     "react-dom": "^16.8.1",
+    "react-transition-group": "^2.5.3",
     "sharp": "^0.21.2"
   },
   "devDependencies": {
     "eslint": "^5.12.0",
-    "nodemon": "^1.18.9"
+    "nodemon": "^1.18.9",
+    "nodemon-webpack-plugin": "^4.0.7",
+    "prettier": "^1.16.4",
+    "webpack": "^4.29.3",
+    "webpack-cli": "^3.2.3"
   }
 }
