+++ public/index.html
@@ -3,21 +3,11 @@
 <head>
   <title>קטמון - העלאת תמונות</title>
   <link rel="stylesheet" href="/stylesheets/style.css">
-  <script src="https://unpkg.com/axios/dist/axios.min.js"></script>
-  <!--
-  For debugging the front-end app-->
-  <script src="/assets/react/umd/react.development.js"></script>
-  <script src="/assets/react-dom/umd/react-dom.development.js"></script>
-  <!--
-  <script src="/assets/react/umd/react.production.js"></script>
-  <script src="/assets/react-dom/umd/react-dom.production.js"></script>
-  -->
-  <script crossorigin src="https://unpkg.com/react-transition-group/dist/react-transition-group.js"></script>
 </head>
 
 <body>
   <div id="app"></div>
-  <script type="module" src="/javascripts/script.js"></script> 
+  <script type="text/javascript" src="/js/index.js"></script> 
 </body>
 
 </html>
