+++ src-client/js/utils/html.js
@@ -1,0 +1,5 @@
+// external
+import htm from 'htm';
+import React from 'react';
+
+export default htm.bind(React.createElement);
