+++ utils/storageService.js
@@ -1,11 +1,7 @@
 const storeLocally = require('./storeLocally');
 const generateStorageFunction = require('./storeMediaPlatform');
 
-const {
-    WMP_DOMAIN,
-    WMP_APPID,
-    WMP_SHARED_SECRET,
-} = process.env;
+const { WMP_DOMAIN, WMP_APPID, WMP_SHARED_SECRET } = process.env;
 
 const isMediaPlatform = WMP_APPID && WMP_DOMAIN && WMP_SHARED_SECRET;
 const isLocal = !isMediaPlatform;
@@ -13,7 +9,11 @@ const localStoragePublicPath = isLocal ? 'uploadedFiles' : null;
 
 function selectStorageFunction() {
     if (isMediaPlatform) {
-        return generateStorageFunction(WMP_DOMAIN, WMP_APPID, WMP_SHARED_SECRET);
+        return generateStorageFunction(
+            WMP_DOMAIN,
+            WMP_APPID,
+            WMP_SHARED_SECRET
+        );
     }
 
     if (isLocal) {
