+++ utils/storeLocally.js
@@ -20,7 +20,7 @@ function createBaseFolderIfNotExists(path) {
                 } else {
                     resolve();
                 }
-            })
+            });
         });
     }
 
@@ -35,13 +35,15 @@ function storeLocally(imageStream, collection, fileName, prefix = '') {
 
         await createBaseFolderIfNotExists(collectionFolder);
 
-        const writeStream = fs.createWriteStream(path.join(collectionFolder, storageFileName));
+        const writeStream = fs.createWriteStream(
+            path.join(collectionFolder, storageFileName)
+        );
 
-        writeStream.on('finish', function () {
+        writeStream.on('finish', function() {
             resolve(`${prefix}/${collection}/${storageFileName}`);
         });
 
-        writeStream.on('error', function (err) {
+        writeStream.on('error', function(err) {
             reject(err);
         });
